finish Diplom 3
